Server/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
pymssql
pyodbc
textdistance
pandas
scikit-learn
numpy